% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerDirection.R
\name{soccerDirection}
\alias{soccerDirection}
\title{Add an arrow showing the direction of play to a soccer pitch ggplot.}
\usage{
soccerDirection(plot, direction = c("right", "left"), lengthPitch = 105,
  widthPitch = 68, arrow_col = "black", grass = FALSE)
}
\arguments{
\item{plot}{an existing ggplot object to add arrow to.}

\item{direction}{character, direction of arrow (\code{"right"} or \code{"left"}).}

\item{lengthPitch, widthPitch}{numeric, length and width of pitch in metres.}

\item{arrow_col}{character, colour of arrow (defaults to \code{"black"}).}

\item{grass}{if TRUE, draws pitch background in green and lines in white. If FALSE, draws pitch background in white and lines in black.}
}
\value{
a ggplot object
}
\description{
Draws an arrow showing the direction of play at the top of an existing soccer pitch ggplot.
}
\examples{
data(tromso)
# draw heatmap of player #9's position
p <- soccerHeatmap(subset(tromso, id == 9), bins = 15, lengthPitch = 105, widthPitch = 68)
# add arrow showing direction of play to the right
soccerDirection(p, "right", lengthPitch = 105, widthPitch = 68)

}
\seealso{
\code{\link{soccerPitchBG}} and \code{\link{soccerPitchFG}} for drawing a soccer pitch
}
