% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerFlipDirection.R
\name{soccerFlipDirection}
\alias{soccerFlipDirection}
\title{Flips x,y-coordinates horizontally in one half to account for changing sides at half-time}
\usage{
soccerFlipDirection(dat, pitchLength = 105, pitchWidth = 68,
  period = "period", periodToFlip = 1, x = "x", y = "y")
}
\arguments{
\item{dat}{= dataframe containing unnormalised x,y-coordinates}

\item{pitchLength, pitchWidth}{= length, width of pitch in metres}

\item{period}{= name of variable containing period labels}

\item{periodToFlip}{= identity of period to flip}

\item{x, y}{= name of variables containing x,y-coordinates}
}
\value{
a dataframe
}
\description{
Normalises direction of attack in both halves of both teams by 
flipping x,y-coordinates horizontally in either the first or second half; 
i.e. teams attack in the same direction all game despite changing sides at 
half-time.
}
\examples{
# fake period data for tromso dataset, and flip direction of '2nd half'
tromso \%>\% 
  mutate(period = if_else(t > as.POSIXct("2013-11-07 21:14:00 GMT"), 1, 2))
  soccerFlipDirection(pitchLength = 120, pitchWidth = 80, periodToFlip = 2)

}
