% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerFlipHoriz.R
\name{soccerFlipHoriz}
\alias{soccerFlipHoriz}
\title{Flips x,y-coordinates horizontally in one half to account for changing sides at half-time}
\usage{
soccerFlipHoriz(dat, periodVar = "period", periodToFlip = 1,
  pitchLength = 105, pitchWidth = 68)
}
\arguments{
\item{dat}{= dataframe containing unnormalised x,y-coordinates named `x` and `y`}

\item{periodVar}{= name of variable containing period labels}

\item{periodToFlip}{= which period to flip}

\item{pitchLength, pitchWidth}{= length, width of pitch in metres}
}
\value{
a dataframe
}
\description{
Normalises direction of attack in both halves of both teams by 
flipping x,y-coordinates horizontally in either the first or second half; 
i.e. teams attack in the same direction all game despite changing sides at 
half-time.
}
\examples{
# to flip coordinates in 2nd half of a dataframe with 1st/2nd half identity labelled by variable named `period`
soccerFlipHoriz(df, "period", 2, 105, 68)

}
