% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerFlow.R
\name{soccerFlow}
\alias{soccerFlow}
\title{Draw a flow field on a soccer pitch.}
\usage{
soccerFlow(df, bins, lengthPitch = 105, widthPitch = 68, yBins = NULL,
  grass = FALSE, plot = NULL)
}
\arguments{
\item{df}{dataframe containing x,y-coordinates of player position in columns named \code{'x'} and \code{'y'}.}

\item{bins}{integer, the number of horizontal bins (length-wise) the soccer pitch is to be divided up into. If no value for \code{yBins} is provided, this value will also be used for the number of vertical (width-wise) bins.}

\item{lengthPitch, widthPitch}{numeric, length and width of pitch in metres.}

\item{yBins}{integer, the number of vertical bins (width-wise) the soccer patch is to be divided up into. If \code{NULL}, the same value is used as for \code{bins}}

\item{grass}{if TRUE, draws pitch background in green and lines in white. If FALSE, draws pitch background in white and lines in black.}

\item{plot}{optional, adds wagon wheels to an existing ggplot object if provided}
}
\value{
a ggplot object of a heatmap on a soccer pitch.
}
\description{
Draws a flow field showing the mean direction of movement made in each sector of the pitch and adds pitch outlines. Note: This function is prototypical and intended to eventually visualise pass and shot event data, but there are no open-source samples of such data available as yet.
}
\examples{
data(tromso_extra)
# draw flow field showing mean direction of player #8's movement
soccerFlow(subset(tromso_extra, id == 8), bins = 5, grass = TRUE)
# draw flow field over player heatmap
p <- soccerHeatmap(subset(tromso_extra, id == 8), bins = 5)
soccerFlow(subset(tromso_extra, id == 8), bins = 5, plot = p)

}
\seealso{
\code{\link{soccerHeatmap}} for drawing a heatmap of player position, or \code{\link{soccerSpokes}} for drawing spokes to show all directions in each area of the pitch.
}
