% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerHeatmap.R
\name{soccerHeatmap}
\alias{soccerHeatmap}
\title{Draw a heatmap on a soccer pitch.}
\usage{
soccerHeatmap(df, bins = 5, lengthPitch = 105, widthPitch = 68,
  yBins = NULL, colLow = "white", colHigh = "red")
}
\arguments{
\item{df}{dataframe containing x,y-coordinates of player position in columns named \code{'x'} and \code{'y'}.}

\item{bins}{integer, the number of horizontal bins (length-wise) the soccer pitch is to be divided up into. If no value for \code{yBins} is provided, this value will also be used for the number of vertical (width-wise) bins.}

\item{lengthPitch, widthPitch}{numeric, length and width of pitch in metres.}

\item{yBins}{integer, the number of vertical bins (width-wise) the soccer patch is to be divided up into. If \code{NULL}, the same value is used as for \code{bins}.}

\item{colLow, colHigh}{character, colours for the low and high ends of the heatmap gradient.}
}
\value{
a ggplot object of a heatmap on a soccer pitch.
}
\description{
Draws a heatmap showing player position frequency in each area of the pitch and adds soccer pitch outlines.
}
\details{
uses \code{ggplot2::geom_bin2d} to map 2D bin counts
}
\examples{
data(tromso)
# draw heatmap of player #9's position
soccerHeatmap(subset(tromso, id == 8), bins = 15)

}
\seealso{
\code{\link{soccerPitchBG}} for a background soccer pitch for the purpose of drawing position maps, player trajectories, etc..
}
