% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerPitch.R
\name{soccerPitch}
\alias{soccerPitch}
\title{Plot a full soccer pitch}
\usage{
soccerPitch(lengthPitch = 105, widthPitch = 68, arrow = c("none", "r",
  "l"), title = NULL, subtitle = NULL, theme = c("light", "dark", "grey",
  "grass"))
}
\arguments{
\item{lengthPitch, widthPitch}{length and width of pitch in metres}

\item{arrow}{optional, adds arrow showing team attack direction as right (\code{'r'}) or left (\code{'l'})}

\item{title, subtitle}{optional, adds title and subtitle to plot}

\item{theme}{draws a \code{light}, \code{dark}, \code{grey}, or \code{grass} coloured pitch}

\item{fillPitch, colPitch}{pitch fill and line colour}
}
\value{
a ggplot object
}
\description{
Draws a soccer pitch as a ggplot object for the purpose of adding layers such as player positions, player trajectories, etc..
}
\examples{
# custom plot of France defensive pressure events vs. Argentina
data(statsbomb)
soccerPitch(arrow = "r", theme = "grass", 
              title = "France (vs. Argentina)", 
              subtitle = "Pressure events") + 
  geom_point(data = filter(statsbomb, team.name == "France" & type.name == "Pressure"), 
             aes(x = location.x, y = location.y), 
             col = "blue", alpha = 0.5)

}
