% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerPitchBG.R
\name{soccerPitchBG}
\alias{soccerPitchBG}
\title{Plot a soccer pitch ggplot object}
\usage{
soccerPitchBG(lengthPitch = 105, widthPitch = 68, fillPitch = "white",
  colPitch = "grey60", grass = FALSE, arrow = c("none", "r", "l"),
  arrow_col = "default", lwd = 0.5, title = NULL, subtitle = NULL)
}
\arguments{
\item{lengthPitch, widthPitch}{length and width of pitch in metres}

\item{fillPitch, colPitch}{pitch fill and line colour}

\item{grass}{if \code{TRUE}, uses a more realistic pitch}

\item{arrow}{optional, adds arrow showing team attack direction as right (\code{'r'}) or left (\code{'l'})}

\item{arrow_col}{colour of attack direction arrow}

\item{lwd}{numeric, pitch line width}

\item{title, subtitle}{optional, adds title and subtitle to plot}
}
\value{
a ggplot object
}
\description{
Draws a soccer pitch as a ggplot object for the purpose of adding layers such as player positions, player trajectories, etc..
}
\examples{
# get x,y-coords of player #8 during first 10 minutes
data(tromso)
dd <- subset(tromso, id == 9)[1:1200,]
# draw player path on pitch
soccerPitchBG(grass = TRUE) + 
  geom_path(data = dd, aes(x, y))

}
\seealso{
\code{\link{soccerPitchFG}} for adding soccer pitch lines to an existing ggplot object
}
