% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerPitchFG.R
\name{soccerPitchFG}
\alias{soccerPitchFG}
\title{Helper function to draw soccer pitch outlines over an existing ggplot object}
\usage{
soccerPitchFG(
  plot,
  lengthPitch = 105,
  widthPitch = 68,
  colPitch = "black",
  arrow = c("none", "r", "l"),
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{plot}{an existing ggplot object to add pitch lines layer to}

\item{lengthPitch, widthPitch}{length and width of pitch in metres}

\item{colPitch}{colour of pitch markings}

\item{arrow}{adds team direction of play arrow as right (\code{'r'}) or left (\code{'l'}); \code{'none'} by default}

\item{title, subtitle}{adds title and subtitle to plot; NULL by default}
}
\value{
a ggplot object
}
\description{
Adds soccer pitch outlines (with transparent fill) to an existing ggplot object (e.g. heatmaps, passing maps, etc..)
}
\seealso{
\code{\link{soccerPitch}} for plotting a soccer pitch as background layer
}
