% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerPositions.R
\name{soccerPositions}
\alias{soccerPositions}
\title{Plot average player position on a soccer pitch.}
\usage{
soccerPositions(df, id_var = "id", lengthPitch = 105, widthPitch = 68,
  col1 = "red", col2 = "white", size = 8, grass = FALSE)
}
\arguments{
\item{df}{dataframe containing x,y-coordinates of player position in columns named \code{'x'} and \code{'y'}.}

\item{id_var}{character, the name of the column containing player identity. Defaults to \code{'id'}.}

\item{lengthPitch, widthPitch}{numeric, length and width of pitch in metres.}

\item{col1}{character, fill colour of position points.}

\item{col2}{character, border colour of position points.}

\item{size}{numeric, size of position points and text.}

\item{grass}{if TRUE, draws pitch background in green and lines in white. If FALSE, draws pitch background in white and lines in black.}
}
\description{
Draws the average x,y-positions of all players in a dataframe and plots over
a soccer pitch.
}
\examples{
data(tromso)
# draw average player position of players
p <- soccerPositions(tromso, lengthPitch = 105, widthPitch = 68, grass = TRUE)
# draw arrow showing direction of play
soccerDirection(p, "right", lengthPitch = 105, widthPitch = 68, grass = TRUE)

}
\seealso{
\code{\link{soccerPitchBG}} for a background soccer pitch for the purpose of drawing position maps, player trajectories, etc..
}
