% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerShortenName.R
\name{soccerShortenName}
\alias{soccerShortenName}
\title{Extract shortened player names}
\usage{
soccerShortenName(names)
}
\arguments{
\item{names}{vector of strings containing full player names}
}
\description{
Helper function to extract last name from full player names
}
\examples{
attach(statsbomb)
statsbomb$name <- soccerShortenName(statsbomb$player.name)

}
