% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerSpokes.R
\name{soccerSpokes}
\alias{soccerSpokes}
\title{Draws a heatmap of player position and plots over soccer pitch outlines.}
\usage{
soccerSpokes(dat, bins, lengthPitch = 105, widthPitch = 68, yBins = NULL,
  heatmap = FALSE)
}
\arguments{
\item{bins}{integer, the number of horizontal bins (length-wise) the soccer pitch is to be divided up into. If no value for \code{yBins} is provided, this value will also be used for the number of vertical (width-wise) bins.}

\item{lengthPitch, widthPitch}{numeric, length and width of pitch in metres.}

\item{yBins}{integer, the number of vertical bins (width-wise) the soccer patch is to be divided up into. If \code{NULL}, the same value is used as for \code{bins}.}

\item{heatmap}{if TRUE, draws a heatmap of player position underneath spokes with same number of bins as spokes}

\item{df}{dataframe containing x,y-coordinates of player position in columns named \code{'x'} and \code{'y'}.}

\item{colLow, colHigh}{character, colours for the low and high ends of the heatmap gradient, if heatmap argument is TRUE}
}
\value{
a ggplot object of a heatmap on a soccer pitch.
}
\description{
Draws a heatmap of player position and plots over soccer pitch outlines.
}
\examples{
\dontrun{
dat <- data(tromso)
# draw heatmap of player #9's position
soccerHeatmap(subset(d, id == 8), bins = 15)
}
}
\seealso{
\code{\link{soccerHeatmap}} for drawing just a heatmap of player position.
}
