% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerStandardiseCols.R
\name{soccerStandardiseCols}
\alias{soccerStandardiseCols}
\title{Rename columns in a dataframe for easier use with other {soccermatics} functions}
\usage{
soccerStandardiseCols(df, method = c("statsbomb"))
}
\arguments{
\item{df}{a dataframe of Statsbomb event data}

\item{method}{source of data; only \code{"statsbomb"} currently supported}
}
\value{
a dataframe with column names x, y, distance, angle, player_id, player_name, team_name, event_name
}
\description{
Rename columns (e.g. \code{"location.x"} -> \code{"x"}, \code{"team.name"} -> \code{"team"}, etc...) to interface directly with other soccermatics functions without having to explicitly define column names as arguments. Currently only supports Statsbomb data.
}
\examples{
library(dplyr)
data(statsbomb)

# transform x,y-coords, standardise column names
my_df <- statsbomb \%>\% 
  soccerTransform(method = 'statsbomb') \%>\% 
  soccerStandardiseCols(method = 'statsbomb')
  
# feed to other functions without defining variables,
# x, y, id,distance, angle, etc...
soccerHeatmap(my_df)

}
