% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerTransformSB.R
\name{soccerTransformSB}
\alias{soccerTransformSB}
\title{Normalises all x,y-coordinate data from StatsBomb data}
\usage{
soccerTransformSB(dat)
}
\arguments{
\item{dat}{dataframe returned from StatsBombR}
}
\value{
a dataframe
}
\description{
Normalises x,y-coordinates from StatsBomb data to metre units for use with soccermatics functions. [StatsBomb pitch coordinates](https://github.com/statsbomb/open-data/blob/master/doc/StatsBomb%20Event%20Data%20Specification%20v1.0.2.pdf) appear to be a 120 x 80 yard pitch, and y-coordinates need to be flipped for use with soccermatics.
}
\examples{
# get data
Comps <- FreeCompetitions()
Matches <- FreeMatches(37)
d <- allinfo(Matches[1])

# transform
d <- soccerTransformSB(d)

}
\seealso{
\code{\link{soccerPitchBG}} readily visualises StatsBomb data with the argument `SB = TRUE`. \code{\link{soccerTransform}} transforms any arbitrary x,y-coordinates to metre units
}
