% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerVelocity.R
\name{soccerVelocity}
\alias{soccerVelocity}
\title{Compute instantaneous distance, speed and direction from x,y-coordinates}
\usage{
soccerVelocity(dat)
}
\arguments{
\item{dat}{dataframe containing unnormalised x,y-coordinates \code{x} and \code{y}, time variable \code{'t'}, and player identifier \code{'id'}}
}
\value{
a dataframe with columns \code{'dist'}, \code{'speed'}, and \code{'direction'} added
}
\description{
Compute instantaneous distance moved (in metres), speed (in metres per second), and direction (in radians) between subsequent frames in a dataframe of x,y-coordinates.
}
\examples{
data(tromso)

# calculate distance, speed, and direction for \code{tromso} dataset
soccerVelocity(tromso)

}
