% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerxGTimeline.R
\name{soccerxGTimeline}
\alias{soccerxGTimeline}
\title{Draw a timeline showing cumulative expected goals (xG) over the course of a match using StatsBomb data.}
\usage{
soccerxGTimeline(dat, homeCol = "red", awayCol = "blue", labels = TRUE,
  y_buffer = 0.3)
}
\arguments{
\item{dat}{a dataframe containing StatsBomb data from one full match}

\item{homeCol, awayCol}{colours of the home and away team, respectively}

\item{labels}{include scoreline and goalscorer labels for goals}

\item{y_buffer}{vertical space to add at the top of the y-axis (a quick and dirty way to ensure text annotations are not cropped).}
}
\value{
a ggplot object
}
\description{
Draw a timeline showing cumulative expected goals (xG, excluding penalties and own goals) by two teams over the course of a match, as well as plotting the scoreline and goalscorer at goal events. Currently only works with StatsBomb data but compatability with other (non-StatsBomb) shot data will be added soon.
}
\examples{
data(statsbomb)

# xG timeline of France vs. Argentina
statsbomb \%>\%
  soccerxGTimeline(homeCol = "blue", awayCol = "lightblue", y_buffer = 0.4)

}
