% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsbomb.R
\docType{data}
\name{statsbomb}
\alias{statsbomb}
\title{Sample StatsBomb event data containing the x,y-locations and identity of players involved in pass events, shot events, defensive actions, and more.}
\format{A dataframe containing 12000 frames of x,y-coordinates and timestamps from 11 players.}
\source{
\href{http://home.ifi.uio.no/paalh/dataset/alfheim/}{ZXY Sport Tracking}
}
\usage{
data(statsbomb)
}
\description{
Sample StatsBomb event data from the France vs. Argentina World Cup 2018 game on the 30th June 2018, made publicly available by StatsBomb \href{https://github.com/statsbomb/open-data}{here}. Data contains 145 variables in total, including x,y-coordinates (\code{location.x}, \code{location.y}). StatsBomb pitch dimensions are 120m long and 80m wide, meaning \code{lengthPitch} should be specified as \code{120} and \code{widthPitch} as \code{80}. Event data for all World Cup games (and other competitions) are accessible via the StatsBombR package available \href{https://github.com/statsbomb/StatsBombR}{here}.
}
\examples{
data(statsbomb)

# heatmap of France defensive pressure vs. Argentina (2018-06-30)
statsbomb \%>\% 
  filter(type.name == "Pressure" & team.name == "France") \%>\% 
  soccerHeatmap(x = "location.x", y = "location.y")

}
\references{
\href{https://github.com/statsbomb/open-data}{StatsBomb Open Data}
}
