% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerFlipDirection.R
\name{soccerFlipDirection}
\alias{soccerFlipDirection}
\title{Flips x,y-coordinates horizontally in one half to account for changing sides at half-time}
\usage{
soccerFlipDirection(df, lengthPitch = 105, widthPitch = 68,
  teamToFlip = NULL, periodToFlip = 1, period = "period", team = "team",
  x = "x", y = "y")
}
\arguments{
\item{df}{= dataframe containing unnormalised x,y-coordinates}

\item{lengthPitch, widthPitch}{= length, width of pitch in metres}

\item{periodToFlip}{= identity of period to flip}

\item{period}{= name of variable containing period labels}

\item{x, y}{= name of variables containing x,y-coordinates}
}
\value{
a dataframe
}
\description{
Normalises direction of attack in both halves of both teams by 
flipping x,y-coordinates horizontally in either the first or second half; 
i.e. teams attack in the same direction all game despite changing sides at 
half-time.
}
\examples{
# fake period data for tromso dataset, and flip direction of '2nd half'
tromso \%>\% 
  mutate(period = if_else(t > as.POSIXct("2013-11-07 21:14:00 GMT"), 1, 2))
  soccerFlipDirection(lengthPitch = 120, widthPitch = 80, periodToFlip = 2)

}
