% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerFlipDirection.R
\name{soccerFlipDirection}
\alias{soccerFlipDirection}
\title{Flips x,y-coordinates horizontally in one half to account for changing sides at half-time}
\usage{
soccerFlipDirection(df, lengthPitch = 105, widthPitch = 68,
  teamToFlip = NULL, periodToFlip = 1:2, team = "team",
  period = "period", x = "x", y = "y")
}
\arguments{
\item{df}{= dataframe containing unnormalised x,y-coordinates}

\item{lengthPitch, widthPitch}{= length, width of pitch in metres}

\item{teamToFlip}{= character, name of team to flip. If \code{NULL}, all x,y-coordinates in \code{df} will be flipped}

\item{periodToFlip}{= integer, period(s) to flip}

\item{team}{= character, name of variables containing x,y-coordinates}

\item{period}{= character, name of variable containing period labels}

\item{x, y}{= character, name of variables containing x,y-coordinates}
}
\value{
a dataframe
}
\description{
Normalises direction of attack in both halves of both teams by 
flipping x,y-coordinates horizontally in either the first or second half; 
i.e. teams attack in the same direction all game despite changing sides at 
half-time.
}
\examples{
library(dplyr)

# flip x,y-coords of France in both halves of statsbomb data
data(statsbomb)
statsbomb \%>\% 
  soccerFlipDirection(lengthPitch = 105, widthPitch = 68, teamToFlip = "France", team = "team.name", x = "location.x", y = "location.y")
  
# flip x,y-coords in 2nd half of Tromso, based on a dummy period variable
data(tromso)
tromso \%>\% 
  mutate(period = if_else(t > as.POSIXct("2013-11-07 21:14:00 GMT"), 1, 2)) \%>\% 
  soccerFlipDirection(lengthPitch = 105, widthPitch = 68, periodToFlip = 2)
  
}
