% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerFlow.R
\name{soccerFlow}
\alias{soccerFlow}
\title{Draw a flow field of passing direction on a soccer pitch}
\usage{
soccerFlow(
  df,
  lengthPitch = 105,
  widthPitch = 68,
  xBins = 5,
  yBins = NULL,
  x = "x",
  y = "y",
  angle = "angle",
  distance = "distance",
  col = "black",
  lwd = 0.5,
  arrow = c("none", "r", "l"),
  title = NULL,
  subtitle = NULL,
  theme = c("light", "dark", "grey", "grass"),
  plot = NULL
)
}
\arguments{
\item{df}{dataframe of event data containing fields of start x,y-coordinates, pass distance, and pass angle}

\item{lengthPitch, widthPitch}{numeric, length and width of pitch in metres.}

\item{xBins, yBins}{integer, the number of horizontal (length-wise) and vertical (width-wise) bins the soccer pitch is to be divided up into; if \code{yBins} is NULL (default), it will take the value of \code{xBins}}

\item{x, y, angle, distance}{names of variables containing pass start x,y-coordinates, angle, and distance}

\item{col}{colour of arrows}

\item{lwd}{thickness of arrow segments}

\item{arrow}{adds team direction of play arrow as right (\code{'r'}) or left (\code{'l'}); \code{'none'} by default}

\item{title, subtitle}{adds title and subtitle to plot; NULL by default}

\item{theme}{palette of pitch background and lines, either \code{light} (default), \code{dark}, \code{grey}, or \code{grass}}

\item{plot}{base plot to add path layer to; NULL by default}
}
\value{
a ggplot object of a heatmap on a soccer pitch
}
\description{
A flow field to show the mean angle and distance of passes in zones of the pitch
}
\examples{
library(dplyr)
data(statsbomb)

# transform x,y-coords, filter only France pass events,
# draw flow field showing mean angle, distance of passes per pitch zone
statsbomb \%>\% 
  soccerTransform(method = 'statsbomb') \%>\% 
  filter(team.name == "France" & type.name == "Pass") \%>\% 
  soccerFlow(xBins=7, yBins=5,
             x="location.x", y="location.y", angle="pass.angle", distance="pass.length")
                   
# transform x,y-coords, standarise column names,
# filter only France pass events
my_df <- statsbomb \%>\% 
  soccerTransform(method = 'statsbomb') \%>\% 
  soccerStandardiseCols(method = 'statsbomb') \%>\%
  filter(team_name == "France" & event_name == "Pass")
  
# overlay flow field onto heatmap showing proportion of team passes per pitch zone
soccerHeatmap(my_df, xBins=7, yBins=5) \%>\%
  soccerFlow(my_df, xBins=7, yBins=5, plot = .)
           
}
\seealso{
\code{\link{soccerHeatmap}} for drawing a heatmap of player position, or \code{\link{soccerSpokes}} for drawing spokes to show all directions in each area of the pitch.
}
