% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerHeatmap.R
\name{soccerHeatmap}
\alias{soccerHeatmap}
\title{Draw a heatmap on a soccer pitch.}
\usage{
soccerHeatmap(df, lengthPitch = 105, widthPitch = 68, xBins = 10,
  yBins = NULL, colPitch = "black", colLow = "white", colHigh = "red")
}
\arguments{
\item{df}{dataframe containing x,y-coordinates of player position in columns named \code{'x'} and \code{'y'}.}

\item{lengthPitch, widthPitch}{numeric, length and width of pitch in metres.}

\item{xBins, yBins}{integer, the number of horizontal (length-wise) and vertical (width-wise) bins the soccer pitch is to be divided up into. If no value for \code{yBins} is provided, it will take the value of \code{xBins}.}

\item{colLow, colHigh}{character, colours for the low and high ends of the heatmap gradient.}
}
\value{
a ggplot object of a heatmap on a soccer pitch.
}
\description{
Draws a heatmap showing player position frequency in each area of the pitch and adds soccer pitch outlines.
}
\details{
uses \code{ggplot2::geom_bin2d} to map 2D bin counts
}
\examples{
data(tromso)
# simple heatmap of player #9's position
soccerHeatmap(subset(tromso, id == 8), xBins = 10)

# draw heatmap with approximately 5m x 5m bins (pitchLength / 5 = 21, pitchWidth / 5 = 13.6)
soccerHeatmap(subset(tromso, id == 8), xBins = 21, yBins = 14)

}
\seealso{
\code{\link{soccerPitchBG}} for a background soccer pitch for the purpose of drawing position maps, player trajectories, etc..
}
