% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerPassmap.R
\name{soccerPassmap}
\alias{soccerPassmap}
\title{Draw a passing network using StatsBomb data}
\usage{
soccerPassmap(df, lengthPitch = 105, widthPitch = 68, minPass = 3,
  fill = "red", col = "black", edgeAlpha = 0.6, edgeCol = NULL,
  label = TRUE, shortNames = TRUE, maxNodeSize = 30, maxEdgeSize = 30,
  labelSize = 4, arrow = c("none", "r", "l"), theme = c("light", "dark",
  "grey", "grass"), title = NULL)
}
\arguments{
\item{df}{dataframe containing x,y-coordinates of player passes}

\item{lengthPitch, widthPitch}{numeric, length and width of pitch, in metres}

\item{fill, col}{fill and border colour of nodes}

\item{edgeAlpha}{transparency of edge lines, from \code{0} - \code{1}. Defaults to \code{0.6} so overlapping edges are visible.}

\item{edgeCol}{colour of edge lines. Default is complementary to \code{theme} colours.}

\item{label}{boolean, draw labels}

\item{shortNames}{shorten player names to display last name as label}

\item{maxNodeSize}{maximum size of nodes}

\item{maxEdgeSize}{maximum width of edge lines}

\item{labelSize}{size of player name labels}

\item{arrow}{optional, adds arrow showing team attack direction as right (\code{'r'}) or left (\code{'l'})}

\item{theme}{draws a \code{light}, \code{dark}, \code{grey}, or \code{grass} coloured pitch}

\item{title}{adds custom title to plot. Defaults to team name.}
}
\description{
Draw an undirected passing network of completed passes on pitch from StatsBomb data. Nodes are scaled by number of successful passes; edge width is scaled by number of successful passes between each node pair. Only passes made until first substition shown (ability to specify custom minutes will be added soon). Total number of passes attempted and percentage of completed passes shown. Compatability with other (non-StatsBomb) shot data will be added soon.
}
\examples{
# France vs. Argentina, minimum of three passes
library(dplyr)
library(soccermatics)

# Argentina pass map until first substituton with transparent edges
statsbomb \%>\% 
  filter(team.name == "Argentina") \%>\% 
  soccerPassmap(fill = "lightblue", arrow = "r",
                title = "Argentina (vs France, 30th June 2018)")

# France pass map until first substituton with opaque edges
statsbomb \%>\% 
  filter(team.name == "France") \%>\% 
  soccerPassmap(fill = "blue", minPass = 3,
                maxEdgeSize = 30, edgeCol = "grey40", edgeAlpha = 1,
                title = "France (vs Argentina, 30th June 2018)")
}
