% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerPath.R
\name{soccerPath}
\alias{soccerPath}
\title{Draw a path of player trajectory on a soccer pitch.}
\usage{
soccerPath(dat, lengthPitch = 105, widthPitch = 68, id_var = NULL,
  col = "black", fillPitch = "white", colPitch = "grey60",
  grass = FALSE, lwd = 1, legend = TRUE, plot = NULL)
}
\arguments{
\item{dat}{dataframe containing x,y-coordinates of player position in columns named \code{'x'} and \code{'y'}}

\item{lengthPitch, widthPitch}{length and width of pitch in metres}

\item{id_var}{character, the name of the column containing player identity. Only required if \code{'dat'} contains multiple players}

\item{col}{colour of path if no \code{'id_var'} is provided. If an \code{'id_var'} is present, colours from ColorBrewer's 'Paired' palette are used}

\item{grass}{if TRUE, draws a more realistic looking pitch}

\item{lwd}{player path thickness}
}
\value{
a ggplot object
}
\description{
Draws a path connecting consecutive x,y-coordinates of a player on a soccer pitch.
}
\examples{
data(tromso)
# draw path of player #8 over first 1200 frames
subset(tromso, id == 8)[1:1200,] \%>\%
  soccerPath(col = "red", grass = TRUE)
# draw path of all players over first 1200 frames
tromso \%>\%
  dplyr::group_by(id) \%>\%
  dplyr::slice(1:1200) \%>\%
  soccerPath("id")

}
