% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerPath.R
\name{soccerPath}
\alias{soccerPath}
\title{Draw a path of player trajectory on a soccer pitch.}
\usage{
soccerPath(df, id_var = NULL, lengthPitch = 105, widthPitch = 68,
  grass = FALSE, col = "black", lwd = 1, legend = TRUE)
}
\arguments{
\item{df}{dataframe containing x,y-coordinates of player position in columns named \code{'x'} and \code{'y'}.}

\item{id_var}{character, the name of the column containing player identity. Only required if \code{'df'} contains multiple players.}

\item{lengthPitch, widthPitch}{numeric, length and width of pitch in metres.}

\item{grass}{if TRUE, draws pitch background in green and lines in white. If FALSE, draws pitch background in white and lines in black.}

\item{col}{colour of path if no \code{'id_var'} is provided. If an \code{'id_var'} is present, colours from ColorBrewer's 'Paired' palette are used.}

\item{lwd}{thickness of path}
}
\value{
a ggplot object
}
\description{
Draws a path connecting consecutive x,y-coordinates of a player on a soccer pitch.
}
\examples{
data(tromso)
# draw path of player #8 over first 1200 frames
subset(tromso, id == 8)[1:1200,] \%>\%
  soccerPath(col = "red", grass = TRUE)
# draw path of all players over first 1200 frames
tromso \%>\%
  dplyr::group_by(id) \%>\%
  dplyr::slice(1:1200) \%>\%
  soccerPath("id")

}
