% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerPath.R
\name{soccerPath}
\alias{soccerPath}
\title{Draw a path of player trajectory on a soccer pitch.}
\usage{
soccerPath(df, lengthPitch = 105, widthPitch = 68, col = "black",
  arrow = c("none", "r", "l"), theme = c("light", "dark", "grey", "grass"),
  lwd = 1, title = NULL, subtitle = NULL, legend = FALSE, x = "x",
  y = "y", id = NULL, plot = NULL)
}
\arguments{
\item{df}{dataframe containing x,y-coordinates of player position}

\item{lengthPitch, widthPitch}{length and width of pitch in metres}

\item{col}{colour of path if no \code{'id'} is provided. If an \code{'id'} is present, colours from ColorBrewer's 'Paired' palette are used}

\item{arrow}{optional, adds arrow showing team attack direction as right (\code{'r'}) or left (\code{'l'})}

\item{theme}{draws a \code{light}, \code{dark}, \code{grey}, or \code{grass} coloured pitch}

\item{lwd}{player path thickness}

\item{title, subtitle}{optional, adds title and subtitle to plot}

\item{legend}{boolean, include legend}

\item{x, y}{= name of variables containing x,y-coordinates}

\item{id}{character, the name of the column containing player identity (only required if \code{'df'} contains multiple players)}

\item{plot}{plot to add path to, if desired}
}
\value{
a ggplot object
}
\description{
Draws a path connecting consecutive x,y-coordinates of a player on a soccer pitch.
}
\examples{
data(tromso)
# draw path of Tromso #8 over first 3 minutes (1800 frames)
subset(tromso, id == 8)[1:1800,] \%>\%
  soccerPath(col = "red", grass = TRUE, arrow = "r")
  
# draw path of all Tromso players over first minute (600 frames)
tromso \%>\%
  dplyr::group_by(id) \%>\%
  dplyr::slice(1:1200) \%>\%
  soccerPath(id = "id")

}
