% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerPitch.R
\name{soccerPitch}
\alias{soccerPitch}
\title{Plot a full soccer pitch}
\usage{
soccerPitch(
  lengthPitch = 105,
  widthPitch = 68,
  arrow = c("none", "r", "l"),
  title = NULL,
  subtitle = NULL,
  theme = c("light", "dark", "grey", "grass"),
  data = NULL
)
}
\arguments{
\item{lengthPitch, widthPitch}{length and width of pitch in metres}

\item{arrow}{adds team direction of play arrow as right (\code{'r'}) or left (\code{'l'}); \code{'none'} by default}

\item{title, subtitle}{adds title and subtitle to plot; NULL by default}

\item{theme}{palette of pitch background and lines, either \code{light} (default), \code{dark}, \code{grey}, or \code{grass}}

\item{data}{a default dataset for plotting in subsequent layers; NULL by default}
}
\value{
a ggplot object
}
\description{
Draws a soccer pitch as a ggplot object for the purpose of adding layers such as player positions, player trajectories, etc..
}
\examples{
library(ggplot2)
data(statsbomb)

# transform Statsbomb coordinates to metre units for plotting
my_df <- soccerTransform(statsbomb, method = "statsbomb")

# filter events of interest (France defensive pressure events vs. Argentina)
my_df <- my_df \%>\% 
  dplyr::filter(team.name == "France" & type.name == "Pressure")

# add custom layers to soccerPitch base
soccerPitch(data = my_df,
            arrow = "r", theme = "grass", 
            title = "France (vs. Argentina)", 
            subtitle = "Pressure events") + 
  geom_point(aes(x = location.x, y = location.y), 
             col = "blue", alpha = 0.5)

}
