% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerPitchBG.R
\name{soccerPitchBG}
\alias{soccerPitchBG}
\title{Draw a soccer pitch.}
\usage{
soccerPitchBG(lengthPitch = 105, widthPitch = 68, fillPitch = "white",
  colPitch = "grey60", grass = FALSE, lwd = 1, border = c(4, 4, 4, 4),
  direction = c("none", "r", "l"), SB = FALSE)
}
\arguments{
\item{lengthPitch, widthPitch}{length and width of pitch in metres}

\item{fillPitch}{pitch fill colour}

\item{colPitch}{pitch line colour}

\item{grass}{if \code{TRUE}, draws a more realistic looking pitch}

\item{lwd}{pitch line width}

\item{border}{size of border drawn around pitch perimeter (t,r,b,l)}
}
\value{
a ggplot object
}
\description{
Draws a soccer pitch as a ggplot object for the purpose of adding layers such as player positions, player trajectories, etc..
}
\examples{
# get x,y-coords of player #8 during first 10 minutes
data(tromso)
dd <- subset(tromso, id == 9)[1:1200,]
# draw player path on pitch
soccerPitchBG(lengthPitch = 105, widthPitch = 68, grass = TRUE) + 
  geom_path(data = dd, aes(x, y), lwd = 1.2)

}
\seealso{
\code{\link{soccerPitchFG}} for drawing a soccer pitch as foreground over an existing ggplot object
}
