% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerPitchBG.R
\name{soccerPitchBG}
\alias{soccerPitchBG}
\title{Draw a soccer pitch.}
\usage{
soccerPitchBG(lengthPitch = 105, widthPitch = 68, grass = FALSE)
}
\arguments{
\item{lengthPitch, widthPitch}{numeric, length and width of pitch in metres.}

\item{grass}{if TRUE, draws pitch background in green and lines in white. If FALSE, draws pitch background in white and lines in black.}
}
\value{
a ggplot object
}
\description{
Draws a soccer pitch as a ggplot object for the purpose of adding layers such as player positions, player trajectories, etc..
}
\examples{
# get x,y-coords of player #8 during first 10 minutes
data(tromso)
dd <- subset(tromso, id == 9)[1:1200,]
# draw player path on pitch
soccerPitchBG(lengthPitch = 105, widthPitch = 68, grass = TRUE) + 
  geom_path(data = dd, aes(x, y), lwd = 2)

}
\seealso{
\code{\link{soccerPitchFG}} for drawing a soccer pitch as foreground over an existing ggplot object
}
