% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerPitchFG.R
\name{soccerPitchFG}
\alias{soccerPitchFG}
\title{Helper function to add soccer pitch outlines to an existing ggplot object}
\usage{
soccerPitchFG(plot, lengthPitch = 105, widthPitch = 68,
  colPitch = "black", arrow = c("none", "r", "l"), arrow_col = "black",
  lwd = 0.5, title = NULL, subtitle = NULL)
}
\arguments{
\item{plot}{an existing ggplot object to add pitch lines to}

\item{lengthPitch, widthPitch}{length and width of pitch in metres}

\item{colPitch}{pitch fill and line colour}

\item{arrow}{optional, adds arrow showing team attack direction as right (\code{'r'}) or left (\code{'l'})}

\item{arrow_col}{colour of attack direction arrow}

\item{lwd}{numeric, pitch line width}

\item{title, subtitle}{optional, adds title and subtitle to plot}
}
\value{
a ggplot object
}
\description{
Adds soccer pitch outlines (with transparent fill) to an existing ggplot object (e.g. heatmaps, passing maps, etc..)
}
\seealso{
\code{\link{soccerPitch}} for plotting a soccer pitch for the purpose of drawing over event data, average position, player trajectories, etc..
}
