% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerPitchFG.R
\name{soccerPitchFG}
\alias{soccerPitchFG}
\title{Add soccer pitch outlines to an existing ggplot}
\usage{
soccerPitchFG(plot, lengthPitch = 105, widthPitch = 68,
  colPitch = "black", lwd = 1)
}
\arguments{
\item{plot}{an existing ggplot object to add layers to.}

\item{lengthPitch, widthPitch}{numeric, length and width of pitch in metres.}
}
\value{
a ggplot object
}
\description{
Draws soccer pitch outlines (with transparent fill) over an existing ggplot object to provide context for heatmaps, passing maps, etc..
}
\examples{
data(tromso)
# draw heatmap of player #9's position
p <- soccerHeatmap(subset(tromso, id == 8), bins = 15, lengthPitch = 105, widthPitch = 68)
# add pitch lines to plot
soccerPitchFG(p, lengthPitch = 105, widthPitch = 68)

}
\seealso{
\code{\link{soccerPitchBG}} for a background soccer pitch for the purpose of drawing position maps, player trajectories, etc..
}
