% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerPitchHalf.R
\name{soccerPitchHalf}
\alias{soccerPitchHalf}
\title{Draws a vertical half soccer pitch for the purpose of plotting shotmaps}
\usage{
soccerPitchHalf(lengthPitch = 105, widthPitch = 68, arrow = c("none", "r",
  "l"), theme = c("light", "dark", "grey", "grass"), title = NULL,
  subtitle = NULL)
}
\arguments{
\item{lengthPitch, widthPitch}{length and width of pitch in metres}

\item{arrow}{optional, adds arrow showing team attack direction as right (\code{'r'}) or left (\code{'l'})}

\item{theme}{draws a \code{light}, \code{dark}, \code{grey}, or \code{grass} coloured pitch}

\item{title, subtitle}{optional, adds title and subtitle to plot}
}
\value{
a ggplot object
}
\description{
Adds soccer pitch outlines (with transparent fill) to an existing ggplot object (e.g. heatmaps, passing maps, etc..)
}
\seealso{
\code{\link{soccerShotmap}} for plotting a shotmap on a half pitch or \code{\link{soccerPitch}} for drawing a full size soccer pitch
}
