% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerResample.R
\name{soccerResample}
\alias{soccerResample}
\title{Resample the frequency of x,y,t- time series with linear interpolation of x,y-coordinates.}
\usage{
soccerResample(dat, r = 10, x = "x", y = "y", z = "z", t = "t")
}
\arguments{
\item{dat}{= dataframe containing x,y-coordinates with time variable}

\item{r}{resampling rate in frames per second}

\item{x, y, z}{= name of variables containing x,y(,z)-coordinates}

\item{t}{= name of variable containing time data}
}
\value{
a dataframe
}
\description{
Downsample or upsample dataframe containing x,y-coordinates and a time variable `t` with linear interpolation of x,y-coordinates and constant interpolation of all other variables.
}
\examples{
# resample tromso dataset from ~21 fps to 10 fps
soccerResample(tromso)

}
