% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerResample.R
\name{soccerResample}
\alias{soccerResample}
\title{Resample the frequency of x,y,t- time series with linear interpolation of x,y-coordinates.}
\usage{
soccerResample(dat, r = 10)
}
\arguments{
\item{dat}{= dataframe containing unnormalised x,y-coordinates named `x` and `y` and a time variable named `t`}

\item{r}{resampling rate in frames per second}
}
\value{
a dataframe
}
\description{
Downsample or upsample dataframe containing x,y-coordinates and a time variable `t` with linear interpolation of x,y-coordinates and constant interpolation of all other variables.
}
\examples{
# resample tromso dataset from ~21 fps to 10 fps
soccerResample(tromso)
}
