% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerShortenName.R
\name{soccerShortenName}
\alias{soccerShortenName}
\title{Extract player surname}
\usage{
soccerShortenName(names)
}
\arguments{
\item{names}{vector of strings containing full player name}
}
\description{
Helper function to extract last name (including common nobiliary particles) from full player names
}
\examples{
data(statsbomb)
statsbomb$name <- soccerShortenName(statsbomb$player.name)

}
