% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerShotmap.R
\name{soccerShotmap}
\alias{soccerShotmap}
\title{Draw a shotmap on a half pitch from StatsBomb data}
\usage{
soccerShotmap(dat, lengthPitch = 105, widthPitch = 68,
  colGoal = "skyblue", colMiss = "grey60", alpha = 0.8, legend = FALSE,
  fillPitch = "white", colPitch = "grey60", grass = FALSE, lwd = 0.5,
  border = c(4, 4, 4, 4), SB = FALSE)
}
\arguments{
\item{lengthPitch, widthPitch}{length and width of pitch in metres}

\item{colGoal, colMiss}{colour of points representing scored and missed shots}

\item{alpha}{transparency of points}

\item{legend}{boolean, include legend or not}

\item{fillPitch}{pitch fill colour}

\item{colPitch}{pitch line colour}

\item{grass}{if \code{TRUE}, draws a more realistic looking pitch}

\item{lwd}{pitch line width}

\item{border}{size of border drawn around pitch perimeter (t,r,b,l)}
}
\value{
a ggplot object
}
\description{
Draw a shotmap on a half pitch from StatsBomb data. Compatability with non-StatsBomb data will be added soon.
}
\examples{
library(StatsBombR)

# get data
Matches <- FreeMatches(37)
d <- allinfo(Matches[1])

# transform all x,y-coordinates of StatsBomb data
d <- soccerTransformSB(d)

# pass map of Manchester City WFC
d \%>\% 
  filter(type.name == "Shot",
  team.name == "Manchester City WFC") \%>\% 
  soccerShotmap(fillPitch = "#1C1F26", colPitch = "white", SB = TRUE)

}
\seealso{
\code{\link{soccerPitchBG}} for drawing a soccer pitch as foreground over an existing ggplot object
}
