% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerShotmap.R
\name{soccerShotmap}
\alias{soccerShotmap}
\title{Draw a shotmap on a half pitch from StatsBomb data}
\usage{
soccerShotmap(dat, lengthPitch = 105, widthPitch = 68,
  colGoal = "skyblue", colMiss = "grey60", alpha = 0.8, legend = FALSE,
  theme = c("light", "dark", "grass"), lwd = 0.5)
}
\arguments{
\item{lengthPitch, widthPitch}{length and width of pitch in metres}

\item{colGoal, colMiss}{colour of points representing scored and missed shots}

\item{alpha}{transparency of points}

\item{legend}{boolean, include legend or not}

\item{lwd}{numeric, pitch line width}

\item{fillPitch, colPitch}{pitch background and line colour}

\item{grass}{if \code{TRUE}, uses a more realistic pitch}
}
\value{
a ggplot object
}
\description{
Draw a shotmap on a half pitch from StatsBomb data. Compatability with other (non-StatsBomb) shot data will be added soon.
}
\examples{
data(statsbomb)

# shot map of France vs. Argentina (2018-06-30)
statsbomb \%>\% 
  filter(team.name == "France") \%>\% 
  soccerShotmap()

}
