% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerSpokes.R
\name{soccerSpokes}
\alias{soccerSpokes}
\title{Draw spokes of passing direction on a soccer pitch}
\usage{
soccerSpokes(
  df,
  lengthPitch = 105,
  widthPitch = 68,
  xBins = 5,
  yBins = NULL,
  angleBins = 8,
  x = "x",
  y = "y",
  angle = "angle",
  minLength = 0.6,
  minAlpha = 0.5,
  minWidth = 0.5,
  col = "black",
  legend = TRUE,
  arrow = c("none", "r", "l"),
  title = NULL,
  subtitle = NULL,
  theme = c("light", "dark", "grey", "grass"),
  plot = NULL
)
}
\arguments{
\item{df}{a dataframe of event data containing fields of start x,y-coordinates, pass distance, and pass angle}

\item{lengthPitch, widthPitch}{numeric, length and width of pitch in metres}

\item{xBins, yBins}{integer, the number of horizontal (length-wise) and vertical (width-wise) bins the soccer pitch is to be divided up into; if \code{yBins} is NULL (default), it will take the value of \code{xBins}}

\item{angleBins}{integer, the number of arrows to draw in each zone of the pitch; for example, a value of 4 clusters has direction vectors up, down, left, and right}

\item{x, y, angle}{names of variables containing pass start x,y-coordinates and angle}

\item{minLength}{numeric, ratio between size of shortest arrow and longest arrow depending on number of events}

\item{minAlpha, minWidth}{numeric, minimum alpha and line width of arrows drawn}

\item{col}{colour of arrows}

\item{legend}{if \code{TRUE}, adds legend for arrow transparency}

\item{arrow}{adds team direction of play arrow as right (\code{'r'}) or left (\code{'l'}); \code{'none'} by default}

\item{title, subtitle}{adds title and subtitle to plot; NULL by default}

\item{theme}{palette of pitch background and lines, either \code{light} (default), \code{dark}, \code{grey}, or \code{grass}}

\item{plot}{base plot to add path layer to; NULL by default}
}
\value{
a ggplot object of a heatmap on a soccer pitch
}
\description{
Multiple arrows to show the distribution of pass angle and distance in zones of the pitch; similar to a radar plot but grouped by pitch location rather than player
}
\examples{
library(dplyr)
data(statsbomb)

# transform x,y-coords, filter only France pass events,
# draw flow field showing mean angle, distance of passes per pitch zone
statsbomb \%>\% 
  soccerTransform(method = 'statsbomb') \%>\% 
  filter(team.name == "France" & type.name == "Pass") \%>\% 
  soccerSpokes(xBins=7, yBins=5, angleBins=12, legend=FALSE)
                   
# transform x,y-coords, standarise column names,
# filter only France pass events
my_df <- statsbomb \%>\%
 soccerTransform(method = 'statsbomb') \%>\%
 soccerStandardiseCols(method = 'statsbomb') \%>\%
  filter(team_name == "France" & event_name == "Pass")
  
# overlay flow field onto heatmap showing proportion of team passes per pitch zone
soccerHeatmap(my_df, xBins=7, yBins=5,
              title = "France passing radar") \%>\%
  soccerSpokes(my_df, xBins=7, yBins=5, angleBins=8, legend=FALSE, plot = .)
           
}
\seealso{
\code{\link{soccerHeatmap}} for drawing a heatmap of player position, or \code{\link{soccerFlow}} for drawing a single arrow for pass distance and angle per pitch zone.
}
