% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerSpokes.R
\name{soccerSpokes}
\alias{soccerSpokes}
\title{Draw spokes on a soccer pitch.}
\usage{
soccerSpokes(df, xBins, lengthPitch = 105, widthPitch = 68,
  angleBins = 16, yBins = NULL, grass = FALSE, legend = TRUE,
  plot = NULL)
}
\arguments{
\item{df}{dataframe containing x,y-coordinates of player position in columns named \code{'x'} and \code{'y'}, and angular information (in radians, ranging between -pi and pi) in a column \code{'direction'}.}

\item{xBins, yBins}{integer, the number of horizontal (length-wise) and vertical (width-wise) bins the soccer pitch is to be divided up into. If no value for \code{yBins} is provided, it will take the value of \code{xBins}.}

\item{lengthPitch, widthPitch}{numeric, length and width of pitch in metres.}

\item{angleBins}{integer, the number of angle bins movement directions are divided up into. For example, a value of 4 clusters directions in each bin into north, east, south and west.}

\item{grass}{if TRUE, draws pitch background in green and lines in white. If FALSE, draws pitch background in white and lines in black.}

\item{legend}{if TRUE, adds legend showing relationship between arrow transparency and number of events}

\item{plot}{optional, adds spokes to an existing ggplot object if provided}
}
\value{
a ggplot object of a heatmap on a soccer pitch.
}
\description{
Draws spokes showing the direction of all movements made in each sector of the pitch. Note: This function is prototypical and intended to eventually visualise pass and shot event data, but there are no open-source samples of such data available as yet.
}
\examples{
data(tromso_extra)
# resample movement dataset to plot 100 movement directions 
# (in absence of pass / shot event data as yet)
id8 <- tromso_extra \%>\%
  filter(id == 8) \%>\%
  sample_n(100)
soccerSpokes(id8, xBins = 5, angleBins = 8, grass = TRUE)
# draw spokes over player heatmap
p <- soccerHeatmap(id8, xBins = 5)
soccerSpokes(id8, xBins = 5, plot = p)

}
\seealso{
\code{\link{soccerHeatmap}} for drawing a heatmap of player position, or \code{\link{soccerSpokes}} for summarising mean direction in each pitch sector
}
