% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerSpokes.R
\name{soccerSpokes}
\alias{soccerSpokes}
\title{Draw spokes on a soccer pitch.}
\usage{
soccerSpokes(df, bins, lengthPitch = 105, widthPitch = 68, yBins = NULL,
  grass = FALSE, plot = NULL)
}
\arguments{
\item{df}{dataframe containing x,y-coordinates of player position in columns named \code{'x'} and \code{'y'}, and angular information of these events in a column \code{'direction'}}

\item{bins}{integer, the number of horizontal bins (length-wise) the soccer pitch is to be divided up into. If no value for \code{yBins} is provided, this value will also be used for the number of vertical (width-wise) bins.}

\item{lengthPitch, widthPitch}{numeric, length and width of pitch in metres.}

\item{yBins}{integer, the number of vertical bins (width-wise) the soccer patch is to be divided up into. If \code{NULL}, the same value is used as for \code{bins}}

\item{grass}{if TRUE, draws pitch background in green and lines in white. If FALSE, draws pitch background in white and lines in black.}

\item{plot}{optional, adds spokes to an existing ggplot object if provided}
}
\value{
a ggplot object of a heatmap on a soccer pitch.
}
\description{
Draws spokes showing the direction of all movements made in each sector of the pitch. Note: This function is prototypical and intended to eventually visualise pass and shot event data, but there are no open-source samples of such data available as yet.
}
\examples{
data(tromso_extra)
# resample movement dataset to plot 100 movement directions (in absence of pass / shot event data as yet)
id8 <- subset(tromso_extra, id == 8)
id8_sample <- id8[sample(1:nrow(id8), 100),]
# draw spokes showing directions of player #8's movement
soccerSpokes(id8_sample, bins = 5, grass = TRUE)
# draw spokes over player heatmap
p <- soccerHeatmap(id8, bins = 5)
soccerSpokes(id8_sample, bins = 5, plot = p)

}
\seealso{
\code{\link{soccerHeatmap}} for drawing a heatmap of player position, or \code{\link{soccerSpokes}} for summarising mean direction in each pitch sector
}
