% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerSpokes.R
\name{soccerSpokes}
\alias{soccerSpokes}
\title{Visualise movement direction on a soccer pitch.}
\usage{
soccerSpokes(plot, df, lengthPitch = 105, widthPitch = 68, xBins,
  yBins = NULL, angleBins = 16, lwd = 0.5, minLength = 0.6,
  minAlpha = 0.4, legend = TRUE)
}
\arguments{
\item{plot}{plot of soccer pitch returned by \code{\link{soccerPitchBG}} to add spokes to}

\item{df}{dataframe containing x,y-coordinates of player position in columns named \code{x} and \code{y} and angular information (in radians, ranging between \code{-pi} and \code{pi}) in a column \code{direction}.}

\item{lengthPitch, widthPitch}{length and width of pitch in metres.}

\item{xBins, yBins}{integer, the number of horizontal (length-wise) and vertical (width-wise) bins the soccer pitch is to be divided up into. If no value for \code{yBins} is provided, it will take the value of \code{xBins}.}

\item{angleBins}{integer, the number of angle bins movement directions are divided up into. For example, a value of 4 clusters directions in each bin into north, east, south and west.}

\item{lwd}{thickness of arrow lines}

\item{minLength}{numeric, ratio between size of shortest arrow and longest arrow depending on number of events.}

\item{legend}{if \code{TRUE}, adds legend showing relationship between arrow transparency and number of events}
}
\value{
a ggplot object
}
\description{
Draws spokes showing the direction of x,y-movements made in each sector of the pitch.
}
\examples{
data(tromso_extra)
# resample movement dataset to plot 100 movement directions 
# (in absence of pass / shot event data as yet)
id8 <- tromso_extra \%>\%
  dplyr::filter(id == 8) \%>\%
  dplyr::sample_n(100)
# 5x5 x,y-bins, 16 angle-bins, blank pitch
soccerPitchBG(pitchLength, pitchWidth) \%>\% 
  soccerSpokes(id8, xBins = 5, angleBins = 16, minLength = 0.4)
# 10x10 x,y-bins, 8 angle-bins, grass pitch
soccerPitchBG(pitchLength, pitchWidth, grass = T) \%>\% 
  soccerSpokes(id8, xBins = 10, angleBins = 8, minLength = 0.2, lwd = 1)
# draw spokes over player heatmap w/ 5x5 x,y-bins, 8 angle-bins
soccerHeatmap(id8, xBins = 5) \%>\% 
  soccerSpokes(id8, xBins = 5, angleBins = 8, lwd = 1)

}
\seealso{
\code{\link{soccerPitchBG}} for drawing a heatmap of player position, \code{\link{soccerHeatmap}} for drawing a heatmap of player position
}
