% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerVelocity.R
\name{soccerVelocity}
\alias{soccerVelocity}
\title{Compute instantaneous distance, speed and direction from x,y-coordinates}
\usage{
soccerVelocity(dat)
}
\arguments{
\item{dat}{dataframe containing unnormalised x,y-coordinates `x` and `y`, time variable `t`, and identifier `id`}
}
\value{
a dataframe
}
\description{
Compute instantaneous distance moved (in metres), speed (in metres per second), and direction (in radians) between subsequent frames in a dataframe of x,y-coordinates.
}
\examples{
# calculate distance, speed, and direction for tromso dataset
soccerVelocity(tromso)

}
